requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
confluent-kafka==1.7.0
dataclasses==0.8
idna==3.3
instaloader==4.8.1
protobuf==3.19.1
pymongo==3.12.1
PyYAML==6.0
requests==2.26.0
urllib3==1.26.7
